# documentation/system-design/book-bazar-openapi.yaml
openapi: 3.0.1
info:
  title: Book Bazar
  description: "[Book Bazar](https://bookbazar.me) API documentation."
  version: 0.0.0
...
servers:
  - url: /api
...
paths:
...
  /user/{id}:
    get:
      tags:
        - User
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Invalid ID supplied
        404:
          description: User was not found
      security:
        - auth: []
    put:
      tags:
        - User
      summary: Update the user with the given ID
      parameters:
        - name: id
          in: path
          description: ID of user to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        description: Updated properties on user object
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ModifiableUser"
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: Invalid ID or request body supplied
        403:
          description: Authenticated user is forbidden from updating the requested user
      security:
        - auth: []
    delete:
      tags:
        - User
      summary: Delete the user with the given ID
      parameters:
        - name: id
          in: path
          description: ID of user to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Successful response
        400:
          description: Invalid ID supplied
        403:
          description: Authenticated user is forbidden from deleting the requested user
      security:
        - auth: []
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        imageUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ModifiableUser:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
          format: binary
...
  securitySchemes:
    auth:
      type: apiKey
      in: cookie
      name: session-token